+++ src/app/locales/en/components/-browser-columns-name/filesystem.js
@@ -1,6 +1,6 @@
 export default {
   name: {
-    files: 'Files',
+    firstColumnName: 'Files',
     size: 'Size',
     owner: 'Owner',
     modification: 'Modified',