+++ src/app/utils/archive-filesystem-browser-model.js
@@ -108,7 +108,22 @@ export default FilesystemBrowserModel.extend({
   /**
    * @override
    */
-  headRowComponentName: 'archive-filesystem-browser/table-head-row',
+  headFirstCellComponentName: 'archive-filesystem-browser/table-head-first-cell',
+
+  /**
+   * @override
+   */
+  headRowTranslation: 'components.archiveFilesystemBrowser.tableHeadRow',
+
+  /**
+   * @override
+   */
+  headRowClass: 'archive-filesystem-table-head-row',
+
+  /**
+   * @override
+   */
+  firstColumnClass: 'fb-table-col-files',
 
   /**
    * @override
@@ -119,6 +134,16 @@ export default FilesystemBrowserModel.extend({
    * @override
    */
   fileFeaturesExtensionComponentName: 'archive-filesystem-browser/file-features-extension',
+
+  /**
+   * @override
+   */
+  isUsingUploadArea: false,
+
+  /**
+   * @override
+   */
+  headRowComponentName: '',
 
   /**
    * @override